+++ include/SpaceFOM/RefFrameDeleted.hh
@@ -21,7 +21,7 @@ NASA, Johnson Space Center\n
 
 @tldh
 @trick_link_dependency{../../source/TrickHLA/Object.cpp}
-@trick_link_dependency{../../source/TrickHLA/ObjectDeleted.cpp}
+@trick_link_dependency{../../source/TrickHLA/ObjectDeletedHandler.cpp}
 @trick_link_dependency{../../source/SpaceFOM/RefFrameDeleted.cpp}
 
 @revs_title
@@ -35,12 +35,12 @@ NASA, Johnson Space Center\n
 #define SPACEFOM_REF_FRAME_DELETED_HH
 
 // TrickHLA includes.
-#include "../TrickHLA/ObjectDeleted.hh"
+#include "../TrickHLA/ObjectDeletedHandler.hh"
 
 namespace SpaceFOM
 {
 
-class RefFrameDeleted : public TrickHLA::ObjectDeleted
+class RefFrameDeleted : public TrickHLA::ObjectDeletedHandler
 {
    // Let the Trick input processor access protected and private data.
    // InputProcessor is really just a marker class (does not really