+++ source/SpaceFOM/DynamicalEntityDeleted.cpp
@@ -17,6 +17,7 @@ NASA, Johnson Space Center\n
 
 @tldh
 @trick_link_dependency{../TrickHLA/Object.cpp}
+@trick_link_dependency{../TrickHLA/ObjectDeletedHandler.cpp}
 @trick_link_dependency{DynamicalEntityDeleted.cpp}
 
 @revs_title